created cent holes and added collsion object with tree squares
Every tile 7 in any map created will now be a collsion object with
uniform height. Right now the code is commented out as the game runs too
slow for true gameplay while calculating collisions with every tree on
the map. Lines 76-79 need to be uncommented to bring this feature back.